refactor as much as possible into separate files
